backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
numpy
scikit-learn
tensorflow
firebase-admin
python-dotenv
aiofiles
